Add files via upload
While the code given works for placing the ships on the board it could really stand to be organized a little better. This is what your task is on the first part of the assignment.

the ShipInfo class has information about all ships but it does nothing for handling the images. It makes sense that there would be a Ship class that holds the general information and specialized ships like minesweep, frigate, battleship, and cruiser are derived from it.

Modify the ShipInfo class so that it is named Ship. Then derive each of the four ship types from Ship. At the very least the derived Ship will hold the images for the ship but you may want to have it hold the images in both directions. You are also going to need a mechanism to return all of the images back. I think it is best to have the class return Labels that hold each of the pieces. This way you are encapsulating the way the ship pieces are created. I will leave it up to you as to how to return a ship but you could return an array of Labels, an ArrayList of Labels just to name a couple.

When you are finished with this step your the program should run the same as it did when it was given to you with the exception that it is now more organized by using inheritance. This means you are going to have to modify the main code for the game to get your new classes to work.
Step 3 A Simple Game

The game is a simple one. To start with you want the ships to be placed around the ocean but you want them hidden. When a player clicks on a Label you should simply replace the image with the image of a hit ship or the image that represents the miss. Your program should keep track of how many misses. When a ship has been sunk you should replace all of the images with the ones that show a sunken ship.
Ship Pieces Hit 	Hit Image
Miss 	Miss Image
Sunk 	Sunk ImageSunk ImageSunk Image End

Please Note that Sunk is going to be several images based on the size of the ship. 

You should also provide a button to reset the game and one to show all ships on the ocean if the user decides to give up.